requirements.txt
gradio==4.25.0
google-generativeai
